Add static helper to extract X509 from cert
Summary:
Since callsites wanting an X509 from an AsyncTransportCertificate will have to
perform a dynamic cast to an OpenSSLTransportCertificiate, we might as well
abstract that away into a static helper.

This makes it potentially a bit easier to change implementation details in the
future without having to update a bunch of existing callsites (and it hides the
ugliness of the dynamic cast).

Reviewed By: yfeldblum, mingtaoy

Differential Revision: D31358948

fbshipit-source-id: 9ed2f81db4cf746a83871961895a0c1f1fb5c297