Immediately cancel any running transform callback on sender cancellation
Summary: Currently, when we get a cancelled notification on the output sender for a transform, we schedule the cancellation of any running transform callback on an executor. CancellationSource is thread-safe, so we don't need to do this. This diff ensures that cancelling the output receiver will immediately trigger cancellation of any running transform callback, allowing the transform callback to deterministically know that it has been cancelled.

Reviewed By: aary

Differential Revision: D32191796

fbshipit-source-id: aabd58be9b58b0f560b8127b336e626c3482a644