Fix crash in MergeChannel when removing subscription to closed receiver
Summary: We currently call closeReceiver() on a receiver in receiversBySubscriptionId_ that has already been removed from receivers_ (and freed). The fix is to check to see whether the receiver in receiversBySubscriptionId_ is still present in receivers_ before invoking any functions on it. If the receiver is not present in receivers_, that means it has already been closed, so we do not need to close it.

Reviewed By: aary

Differential Revision: D32233062

fbshipit-source-id: 0aa763b509897db7de970e12e2f01cffc1b49fc3