don't clone headroom and tailroom when unsharing non-managed input
Summary:
`clone()` preserves headroom and tailroom, so what is happening right
now when deserializing from non-managed buffer (such as wrapped range), is that
every lazy field would result in allocation equivalent in size to the input
buffer!

Apply that to a container and that becomes a combinatorial explosion.

(Note: this ignores all push blocking failures!)

Reviewed By: ot

Differential Revision: D32942233

fbshipit-source-id: b81d8389178b27068dce85b414883d75b1f50ca3