let AsyncBase not use PATH_MAX
Summary:
We may both improve portability and simplify impleementation by using `std::filesystem` in this case, which is just constructing a path and reading the corresponding symlink.

Pure runtime performance is not critical here - this codepath is mostly used for logging - so the extra allocations are not really a problem.

Reviewed By: Gownta

Differential Revision: D32942561

fbshipit-source-id: 6d24b9ba900dcfd8653a843cce539627f7f57735