split smart exception tracer read + register code
Summary:
Split the SmartExceptionTracer so that code can depend on the reading
functionality without also registering the throw hooks at the same time. This
is because hooks during throwing an exception can be expensive for some users.
Splitting it this way allows a library to try to print an async trace if it's
available, but still letting the end application decide if it wants to enable
the feature or not.

Reviewed By: yfeldblum, ot

Differential Revision: D32766294

fbshipit-source-id: 1ffd86d026fc2c18dfa547cf2d169f6491a877cc