OpenSSLPtrTypes: Fix build with OpenSSL 3
Summary:
The history of `sk_*_free` and `sk_*_pop_free`:

* OpenSSL 1.0.2 - These were macros
* OpenSSL 1.1.0 - These changed to static inline functions
* OpenSSL 1.1.1 - These remained as static inline functions (for ABI compatibility with OpenSSL 1.1.0)
* OpenSSL 3.0.0 - These changed to macros.

Give up on trying to use the public interface for deleting `STACK_OF(T)`. This diff switches the implementation
to just invoke the underlying `OPENSSL_sk_free` and `OPENSSL_sk_pop_free`.

If a future version of OpenSSL changes the implementation such that `STACK_OF(T)` is not simply an alias of `OPENSSL_STACK`, we'll deal with it then.

Reviewed By: yfeldblum

Differential Revision: D33071607

fbshipit-source-id: ef37c984a120652c5224410cbafa634d4dc763d1