avoid tail calls in the stack-trace test example
Summary: The test looks for every function in its list of functions to be present in the example stack-trace. But, depending on optimizations, the compiler may transform some of the calls to tail calls, in which case the calling functions would be elided from the stack-trace. Prevent those tail calls.

Reviewed By: kennyyu

Differential Revision: D33797857

fbshipit-source-id: 1c20c19bc612cebbc6a30466ef8deb54713b1497