atomic_fetch_flip
Summary:
The flip operation, like the set and reset operations, are optimizable on x86-related architectures using the `btc` instruction, compared with `bts` and `btr` for set and reset operations.

The optimization is done for gnu-like compilers. However, for MSVC, there is no interlocked intrinsic for `btc`, so the optimization is not done.

Reviewed By: luciang

Differential Revision: D32049844

fbshipit-source-id: 5f60f8c4c534fb0ee41d76c5d766f96f7ff94892