Add a formatter for fbstring
Summary: {fmt} is moving away from relying on implicit conversions because they may introduce ODR violations and not supported consistently. For example, `fbstring` can be formatted via an implicit conversion or via `operator<<`. To make sure this doesn't happen and make the code compatible with future versions of {fmt} add a `formatter` specialization for `fbstring`. This is consistent with `folly::StringPiece` which already has such specialization.

Reviewed By: ot

Differential Revision: D33994222

fbshipit-source-id: 8150415ef45bd97b24aa12dd209b56c977c148c6