simplify Function is-small check
Summary:
In a way which also happens to improve build speed.

It turns out that the bulk of the improvement comes from removing the use of `std::is_nothrow_move_constructible` since it may internally have a `std::declval` in there. But that's not the only source of improvements.

```name=trunk
$ foundation/scripts/run-perf-compile --syntax folly/futures/test/FutureTest.cpp
     8,367,373,932      instructions:uP
```
```name=branch
$ foundation/scripts/run-perf-compile --syntax folly/futures/test/FutureTest.cpp
     8,101,850,897      instructions:uP
```

Reviewed By: luciang

Differential Revision: D32973629

fbshipit-source-id: 37c27763a4f622750719db11fee7cb23c0b0291e