fix missing constexpr specifier in Latch
Summary:
For the `Latch` constructor to be constexpr, at least one path through it must be constexpr, at least even if just in theory. But every path goes through `terminate_if`, which was not `constexpr`. Fix that.

Note: `void`-returning `constexpr` functions is new in C++17. Some compilers accept the code as-is but MSVC correctly rejects it.

Differential Revision: D34334488

fbshipit-source-id: 6f22ce6230b19a74068fd1ae8253d89583befaf2