Change folly::dynamic Type::DOUBLE to use integer hashing
Summary:
Currently folly::dynamic(2).hash() is not equal to folly::dynamic(2.0).hash() - even though operator= will return true between the dynamics given the implicit conversion logic in the code.

While the hash of C++ doubles and ints does not necessarily correlate, even when numerically equal on implicit conversion, the C++ equivalents do not allow the same gotchas with regards hash-based containers.

In C++ doubles and ints cannot co-exist in the same hash-based container given that the container will only have one type - so lookups with a different type don't really make sense. Even if lookup was done for an int into a container of type double - the implicit conversion means that the call of hash will execute on the double, thus the hash will be equal.

With folly::dynamic though - the behavior prior to this will deviate from the C++ mechanics. Lookups into a hash based container of dynamics which are all of type double using an int type will not find (usually) the numerically equivalent double except in the rare case of a hash collision. What's more - usually both an int and its corresponding equivalent double can co-exist in the hash container - but again only when no collisions.

While concurrent usage of ints and doubles in hash maps/sets is unlikely, making the hash behavior of doubles and ints equal will bring consistency between comparison operators and hash, and bring them more in line with what the behavior would be when doing lookups into a hash-based container of doubles.

How this is done is by using the int64_t hashing when double is exactly representable as an integer and within int64_t range. This avoids reducing the hash range when people use e.g. a lot of decimals between 0 and 1. Making the modification to doubles is likely less disruptive than the reverse given the utter rarity of hash containers keyed on doubles compared to int's.

Reviewed By: Gownta

Differential Revision: D33496848

fbshipit-source-id: 1cdad87690a14d068fb3f582b2ab6f0a49521865