use new carry-flag out in atomic-fetch-bit-op asm
Summary:
The `bt` family of x86 instructions load the tested bit into the carry flag. As of gcc6, this carry flag can be marked explicitly as an output of the inline `asm` text, which lets us avoid having to copy the carry flag to a normal register only to test the value of that register again.

Instead of:
```
lock bts
setb
test
jnz
```

We can have:
```
lock bts
jb
```

https://developers.redhat.com/blog/2016/02/25/new-asm-flags-feature-for-x86-in-gcc-6

Reviewed By: ot, luciang

Differential Revision: D32258102

fbshipit-source-id: f8fe028d887ec85035e6e9593adb098fc09c890e