Avoid creating BrokenPromise exceptions during transforms
Summary:
Currently, when a value from a channel is being transformed, we will sometimes destroy a promise without resolving it first. This causes a BrokenPromise exception to be created with a custom-formatted error message, which wastes cycles.

This diff ensures that we always resolve our promise before destroying it. (It also switches from folly::Promise to folly::coro::Promise, which is lighter-weight.)

Reviewed By: Gownta

Differential Revision: D33628778

fbshipit-source-id: b1bff585ac7b9a028c16bf658abe750d7e65134f