Replace clock_gettime with folly_clock_gettime
The clock_gettime situation on Apple platforms is complex enough as it is.
Linking an incomplete implementation in replacement of the global symbol breaks
other libraries including libc++ which uses clock_gettime internally [1].
Depending on the approach taken by the libc++ the current implementation could
trigger infinite recursion calls as well.

Since the goal of the /portability/ is to make Folly portable, we should adapt
and use an approach which is compatible with Apple SDK development while
maintaining compatibility on other platforms.

This patch does that by adding a folly_ prefix to clock functions to avoid
clashing with the native symbol and #define clock_getX folly_clock_getX in order
to maintain source compatibility.

As a side effect, this also addresses the redefinition issues that this [2]
other proposed fix solved.

[1] https://github.com/facebook/folly/issues/1602
[2] https://github.com/facebook/folly/pull/1724